// src/app/shared/document-search/usage-type.enum.ts
export enum UsageType {
    // Otro, diferente al resto de los especificados
    Other = 0,
    // Imprimir copia del documento
    PrintCopy = 1,
    // Rectificar/anular un documento
    Rectify = 2,
    // Pagar documentos pendientes
    PayPending = 3,
    // Anulacion Parcial
    AnulacionParcial = 4,
    // Pago Deduda Masiva
    PayPendingMassive = 5
}